requirements.txt
playwright>=1.49.0
pytest>=8.2.0
pytest-playwright>=0.5.0
allure-pytest>=2.13.5